// web/src/client/types.gen.ts
export type EnvVar = {
  /**
   * Name
   */
  name: string;
  /**
   * Value
   */
  value: string;
};
// ...
export type ServerInTask = {
  /**
   * Name
   * 服务器名称
   */
  name: string;
  /**
   * Ip
   * IP地址
   */
  ip: string;
  /**
   * Sn
   * 序列号
   */
  sn?: string | null;
  /**
   * Id
   * 资源ID
   */
  id: number;
  /**
   * Bmc Ip
   * BMC IP地址
   */
  bmc_ip: string | null;
  /**
   * Bmc Username
   * BMC 用户名
   */
  bmc_username: string | null;
  /**
   * Port
   * 服务端口
   */
  port: number | null;
  /**
   * Username
   * 用户名
   */
  username: string | null;
};
// ...
export type TaskConfig = {
  /**
   * Name
   */
  name: string;
  /**
   * Description
   */
  description?: string | null;
  /**
   * Env Vars
   */
  env_vars?: Array<EnvVar> | null;
  /**
   * Id
   */
  id: number;
  /**
   * Project Id
   */
  project_id: number;
  /**
   * Created At
   */
  created_at: Date;
  /**
   * Updated At
   */
  updated_at: Date;
};
// ...
export type TaskStatus =
  | 'Created'
  | 'Running'
  | 'Passed'
  | 'Failed'
  | 'Canceled'
  | 'Error';

/**
 * TaskStatus
 */
export const TaskStatus = {
  CREATED: 'Created',
  RUNNING: 'Running',
  PASSED: 'Passed',
  FAILED: 'Failed',
  CANCELED: 'Canceled',
  ERROR: 'Error',
} as const;
// ...
export type TestTaskListItem = {
  /**
   * Id
   */
  id: number;
  /**
   * Name
   */
  name: string;
  owner: UserSearch;
  config?: TaskConfig | null;
  status?: TaskStatus | null;
  server?: ServerInTask | null;
  /**
   * Failed Continue
   */
  failed_continue?: number;
  /**
   * Lark Notice
   */
  lark_notice?: number;
  /**
   * Created At
   */
  created_at: Date;
  /**
   * Updated At
   */
  updated_at: Date;
};
// ...
export type UserSearch = {
  /**
   * Id
   */
  id: number;
  /**
   * Email
   */
  email: string;
  /**
   * Name
   */
  name: string;
  /**
   * En Name
   */
  en_name?: string | null;
  /**
   * Union Id
   */
  union_id?: string | null;
  /**
   * Avatar
   */
  avatar?: string | null;
};